refactor(streamlit): unify HC and IA on st.navigation()
Brings the Health Check and Impact Analyzer apps in line with the
Decision Analyzer pattern (PR #733). All three apps now use
st.navigation() instead of the legacy filesystem-based pages/
directory auto-discovery. No user-visible change — same pages,
same URLs, same sidebar entries.

Lays the groundwork for a future cross-app launcher that can host
all three tools' page groups in one Streamlit process while keeping
each tool's sidebar focused.
